nyota-enterprise/security/requirements.txt
nats-py
pydantic
psycopg2-binary
